ajustes en capa domain ubigeo

+++ ubigeo.microservice/src/main/java/com/nsrtm/ubigeo/microservice/domain/Auditoria.java
@@ -7,14 +7,14 @@ import java.util.Date;
 @MappedSuperclass
 public class Auditoria {
     @Column(name = "USUARIO_CREACION")
-    public Integer usuarioCreacion;
+    public Long usuarioCreacion;
     @Column(name = "FECHA_CREACION")
     public Date fechaCreacion;
 
     @Column(name = "TERMINAL_CREACION")
     public String terminalCreacion;
     @Column(name = "USUARIO_MODIFICACION")
-    public Integer usuarioEdicion;
+    public Long usuarioEdicion;
     @Column(name = "FECHA_MODIFICACION")
     public Date fechaModificacion;
